clauditor-24h.9: US-009 — extract grade into cli/grade.py with sub-helpers
Move cmd_grade, _cmd_grade_with_workspace, _run_baseline_phase,
_find_prior_grading_json, _print_grade_diff, and
_print_baseline_delta_block out of cli/__init__.py into a new
cli/grade.py module. Decompose _cmd_grade_with_workspace into
single-purpose helpers so no orchestration function exceeds 100 lines.

New helpers in cli/grade.py:
- _load_and_validate_grade_args: pre-LLM input validation
- _apply_only_criterion_filter: in-place criterion filter
- _run_skill_variants: primary + --variance skill runs
- _grade_all_runs: concurrent LLM grading over run outputs
- _build_variance_report: aggregate per-run stability stats
- _build_grade_metrics: bucketed tokens/duration aggregation
- _write_workspace_sidecars: stage grading/assertions/extraction/baseline
- _write_run_dirs_or_scrub, _write_assertions_sidecar,
  _write_extraction_sidecar, _write_baseline_and_benchmark: sub-steps
- _write_timing_and_finalize: timing.json + atomic rename
- _report_grade_to_user: bundled post-finalize reporting
- _print_grade_output, _emit_diff_and_baseline_blocks,
  _append_grade_history_record, _check_min_baseline_delta_gate

cli/__init__.py shrinks by 840 lines. cmd_grade and _run_baseline_phase
are re-exported so existing test imports (from clauditor.cli import
cmd_grade) continue to work. Help output is byte-identical; all 1236
tests pass; cli/grade.py coverage is 98%.
